ipc-do-cyclic-id-allocation-for-the-ipc-object-fix
fix max() warning

  ipc/util.c: In function 'ipc_idr_alloc':
   include/linux/kernel.h:828:29: warning: comparison of distinct pointer types lacks a cast
      (!!(sizeof((typeof(x) *)1 == (typeof(y) *)1)))

Reported-by: kbuild test robot <lkp@intel.com>
Cc: Manfred Spraul <manfred@colorfullife.com>

+++ ipc/util.h
@@ -42,7 +42,7 @@ extern int ipc_min_cycle;
 #else /* CONFIG_SYSVIPC_SYSCTL */
 
 #define ipc_mni			IPCMNI
-#define ipc_min_cycle		RADIX_TREE_MAP_SIZE
+#define ipc_min_cycle		((int)RADIX_TREE_MAP_SIZE)
 #define ipcmni_seq_shift()	IPCMNI_SHIFT
 #define IPCMNI_IDX_MASK		((1 << IPCMNI_SHIFT) - 1)
 #endif /* CONFIG_SYSVIPC_SYSCTL */
